VOTE-332 Move json files to language code folder
