+++ src/components/FormSections/Identification.jsx
@@ -1,6 +1,6 @@
 import { Label, TextInput, Button, Fieldset, Dropdown } from '@trussworks/react-uswds';
 import React, { useState } from "react";
-import content from "../../data/registration-form.json";
+import content from "../../data/en/registration-form.json";
 import { focusNext, restrictLength, restrictType, checkExpiration, checkForErrors } from '../HelperFunctions/ValidateField';
 
 function Identification(props){
@@ -10,8 +10,8 @@ function Identification(props){
     const idNumReq = stateFieldRequirements.ID_num;
     const idNumVisible = stateFieldVisible.ID_num;
 
-    const [handleErrors, setHandleErrors] = useState({ 
-        id_selection: false, 
+    const [handleErrors, setHandleErrors] = useState({
+        id_selection: false,
         id_number: false,
         issue_date: false,
         expire_date: false,
@@ -32,7 +32,7 @@ function Identification(props){
             props.fieldData.id_expire_date_year.length === 4,
             checkExpiration(props.fieldData.id_expire_date_year + "-" + props.fieldData.id_expire_date_month + "-" + props.fieldData.id_expire_date_day)
         ]
-        
+
         if (dateValues.includes(false)) {
             setHandleErrors({ ...handleErrors, [dateType]: (true) })
         } else {
@@ -59,7 +59,7 @@ function Identification(props){
                 <p>{stateInstructions.ID_num_text}</p>
 
                 <div className={(idNumReq && handleErrors.id_selection) ? 'error-container' : ''}>
-                <Dropdown 
+                <Dropdown
                 id="id-num-dropdown"
                 name="input-dropdown"
                 value={props.idType}
@@ -71,29 +71,29 @@ function Identification(props){
                     <option key="id-num" value="state-id-num">State ID Number</option>
                     <option key="ssn" value="ssn">Social Security Number (last 4 digits)</option>
                     <option key="id-none" value="none">I do not have a valid ID number</option>
-                </Dropdown> 
-                {(idNumReq && handleErrors.id_selection) && 
+                </Dropdown>
+                {(idNumReq && handleErrors.id_selection) &&
                     <span id="id-num-dropdown-error" role="alert" className='error-text'>
                         Identification selection must be made from the dropdown.
                     </span>
                 }
-                </div> 
+                </div>
 
-                {props.idType === 'state-id-num' && 
+                {props.idType === 'state-id-num' &&
                 <div>
                 <div className={(idNumReq && handleErrors.id_number) ? 'error-container' : ''}>
                     <Label htmlFor="state-id-num-error">State ID Number{idNumReq && <span className='required-text'>*</span>}
-                    <TextInput 
-                        id="state-id-num" 
-                        name="state-id-num" 
-                        type="text" 
-                        autoComplete="off" 
+                    <TextInput
+                        id="state-id-num"
+                        name="state-id-num"
+                        type="text"
+                        autoComplete="off"
                         required={idNumReq}
-                        value={props.fieldData.id_number} 
-                        onChange={props.saveFieldData('id_number')} 
+                        value={props.fieldData.id_number}
+                        onChange={props.saveFieldData('id_number')}
                         onBlur={(e) => setHandleErrors({ ...handleErrors, id_number: checkForErrors(e, 'check value exists') })}
                     />
-                    {(idNumReq && handleErrors.id_number) && 
+                    {(idNumReq && handleErrors.id_number) &&
                         <span id="state-id-num-error" role="alert" className='error-text'>
                             ID number must be filled out.
                         </span>
@@ -106,12 +106,12 @@ function Identification(props){
                     <span className="usa-hint" id="id-issue-date-hint">
                     For example: January 19 2000
                     </span>
-                    <div 
-                        id="id-issue-date" 
-                        className="usa-memorable-date" 
-                        name="date-of-birth" 
-                        autoComplete="off" 
-                        required={idNumReq} 
+                    <div
+                        id="id-issue-date"
+                        className="usa-memorable-date"
+                        name="date-of-birth"
+                        autoComplete="off"
+                        required={idNumReq}
                         onBlur={event => { if (!event.currentTarget.contains(event.relatedTarget)) checkDateValues('issue'); }}
                         data-testid="dateInputGroup"
                     >
@@ -123,8 +123,8 @@ function Identification(props){
                                     id="id_issue_date_month"
                                     name="id_issue_date_month"
                                     aria-describedby="issue-date-error"
-                                    required={true} 
-                                    value={props.fieldData.id_issue_date_month} 
+                                    required={true}
+                                    value={props.fieldData.id_issue_date_month}
                                     onInput={props.saveFieldData('id_issue_date_month')}
                                     onChange={(e) => {focusNext(e, "id_issue_date_day", "month")}}
                                 >
@@ -147,21 +147,21 @@ function Identification(props){
                             <div data-testid="formGroup" className="usa-form-group usa-form-group--month">
                                 <label data-testid="label" className="usa-label" htmlFor="testDateInput">
                                     Day
-                                <input 
-                                    id="id_issue_date_day" 
-                                    className="usa-input" 
-                                    name="id_issue_date_day" 
+                                <input
+                                    id="id_issue_date_day"
+                                    className="usa-input"
+                                    name="id_issue_date_day"
                                     aria-describedby="issue-date-error"
-                                    label="Day" 
+                                    label="Day"
                                     unit="day"
-                                    required={true} 
-                                    type="number" 
-                                    inputMode="numeric" 
-                                    min={1} 
-                                    max={31} 
-                                    minLength={2} 
+                                    required={true}
+                                    type="number"
+                                    inputMode="numeric"
+                                    min={1}
+                                    max={31}
+                                    minLength={2}
                                     maxLength={2}
-                                    value={props.fieldData.id_issue_date_day} 
+                                    value={props.fieldData.id_issue_date_day}
                                     onInput={props.saveFieldData('id_issue_date_day')}
                                     onChange={(e) => {focusNext(e, "id_issue_date_year"), restrictLength(e, e.target.value, e.target.maxLength) }}
                               />
@@ -170,19 +170,19 @@ function Identification(props){
                             <div data-testid="formGroup" className="usa-form-group usa-form-group--month">
                                 <label data-testid="label" className="usa-label" htmlFor="testDateInput">
                                     Year
-                                <input 
-                                    id="id_issue_date_year" 
-                                    className="usa-input" 
-                                    name="id_issue_date_year" 
+                                <input
+                                    id="id_issue_date_year"
+                                    className="usa-input"
+                                    name="id_issue_date_year"
                                     aria-describedby="issue-date-error"
-                                    label="Year" 
+                                    label="Year"
                                     unit="year"
-                                    required={true} 
-                                    type="text" 
-                                    inputMode="numeric" 
-                                    minLength={4} 
+                                    required={true}
+                                    type="text"
+                                    inputMode="numeric"
+                                    minLength={4}
                                     maxLength={4}
-                                    value={props.fieldData.id_issue_date_year} 
+                                    value={props.fieldData.id_issue_date_year}
                                     onInput={props.saveFieldData('id_issue_date_year')}
                                     onKeyDown={(e) => restrictType(e, 'number')}
                                 />
@@ -190,7 +190,7 @@ function Identification(props){
                             </div>
                     </div>
                 </Fieldset>
-                {(idNumReq && handleErrors.issue_date) && 
+                {(idNumReq && handleErrors.issue_date) &&
                     <span id="issue-date-error" role="alert" className='error-text'>
                     Issue Date must follow the format of January 19 2000.
                     </span>
@@ -202,12 +202,12 @@ function Identification(props){
                     <span className="usa-hint" id="id-issue-date-hint">
                     For example: January 19 2000
                     </span>
-                    <div 
-                        id="id-expire-date" 
-                        className="usa-memorable-date" 
-                        name="date-of-birth" 
-                        autoComplete="off" 
-                        required={idNumReq} 
+                    <div
+                        id="id-expire-date"
+                        className="usa-memorable-date"
+                        name="date-of-birth"
+                        autoComplete="off"
+                        required={idNumReq}
                         onBlur={event => { if (!event.currentTarget.contains(event.relatedTarget)) checkDateValues('expire'); }}
                         data-testid="dateInputGroup"
                     >
@@ -219,8 +219,8 @@ function Identification(props){
                                     id="id_expire_date_month"
                                     name="id_expire_date_month"
                                     aria-describedby="expire-date-error"
-                                    required={true} 
-                                    value={props.fieldData.id_expire_date_month} 
+                                    required={true}
+                                    value={props.fieldData.id_expire_date_month}
                                     onInput={props.saveFieldData('id_expire_date_month')}
                                     onChange={(e) => {focusNext(e, "id_expire_date_day", "month")}}
                                 >
@@ -244,21 +244,21 @@ function Identification(props){
                             <div data-testid="formGroup" className="usa-form-group usa-form-group--month">
                             <label data-testid="label" className="usa-label" htmlFor="testDateInput">
                                 Day
-                                <input 
-                                    id="id_expire_date_day" 
-                                    className="usa-input" 
-                                    name="id_expire_date_day" 
+                                <input
+                                    id="id_expire_date_day"
+                                    className="usa-input"
+                                    name="id_expire_date_day"
                                     aria-describedby="expire-date-error"
-                                    label="Day" 
+                                    label="Day"
                                     unit="day"
-                                    required={true} 
-                                    type="number" 
-                                    inputMode="numeric" 
-                                    min={1} 
-                                    max={31} 
-                                    minLength={2} 
+                                    required={true}
+                                    type="number"
+                                    inputMode="numeric"
+                                    min={1}
+                                    max={31}
+                                    minLength={2}
                                     maxLength={2}
-                                    value={props.fieldData.id_expire_date_day} 
+                                    value={props.fieldData.id_expire_date_day}
                                     onInput={props.saveFieldData('id_expire_date_day')}
                                     onChange={(e) => {focusNext(e, "id_expire_date_year"), restrictLength(e, e.target.value, e.target.maxLength) }}
                                 />
@@ -268,18 +268,18 @@ function Identification(props){
                             <div data-testid="formGroup" className="usa-form-group usa-form-group--month">
                                 <label data-testid="label" className="usa-label" htmlFor="testDateInput">
                                     Year
-                                <input 
-                                    id="id_expire_date_year" 
-                                    className="usa-input" 
-                                    name="id_expire_date_year" 
+                                <input
+                                    id="id_expire_date_year"
+                                    className="usa-input"
+                                    name="id_expire_date_year"
                                     aria-describedby="expire-date-error"
                                     label="Year" unit="year"
-                                    required={true} 
-                                    type="text" 
-                                    inputMode="numeric" 
-                                    minLength={4} 
+                                    required={true}
+                                    type="text"
+                                    inputMode="numeric"
+                                    minLength={4}
                                     maxLength={4}
-                                    value={props.fieldData.id_expire_date_year} 
+                                    value={props.fieldData.id_expire_date_year}
                                     onInput={props.saveFieldData('id_expire_date_year')}
                                     onKeyDown={(e) => restrictType(e, 'number')}
                                 />
@@ -287,7 +287,7 @@ function Identification(props){
                             </div>
                     </div>
                 </Fieldset>
-                {(idNumReq && handleErrors.expire_date) && 
+                {(idNumReq && handleErrors.expire_date) &&
                     <span id="expire-date-error" role="alert" className='error-text'>
                     Expire Date must follow the format of January 19 2000 and be in the future.
                     </span>
@@ -295,24 +295,24 @@ function Identification(props){
                 </div>
                 }
 
-                {props.idType === 'ssn' && 
+                {props.idType === 'ssn' &&
                 <div className={(idNumReq && handleErrors.id_ssn) ? 'error-container' : ''}>
                 <Label htmlFor="ssn-input-error">Social Security Number (last 4 digits){idNumReq && <span className='required-text'>*</span>}
-                <TextInput 
-                    id="ssn-input" 
-                    name="ssn-input" 
-                    autoComplete="off" 
+                <TextInput
+                    id="ssn-input"
+                    name="ssn-input"
+                    autoComplete="off"
                     required={idNumReq}
-                    type="text" 
+                    type="text"
                     inputMode="numeric"
-                    minLength={4} 
-                    maxLength={4} 
-                    value={props.fieldData.id_number} 
+                    minLength={4}
+                    maxLength={4}
+                    value={props.fieldData.id_number}
                     onChange={props.saveFieldData('id_number')}
-                    onKeyDown={(e) => restrictType(e, 'number')} 
+                    onKeyDown={(e) => restrictType(e, 'number')}
                     onBlur={(e) => setHandleErrors({ ...handleErrors, id_ssn: checkForErrors(e, 'check value length') })}
                     />
-                    {(idNumReq && handleErrors.id_ssn) && 
+                    {(idNumReq && handleErrors.id_ssn) &&
                     <span id="ssn-input-error" role="alert" className='error-text'>
                         Social Security Number must be 4 digits.
                     </span>
