+++ src/components/PathSelection.jsx
@@ -1,5 +1,5 @@
 import { Button, Grid } from '@trussworks/react-uswds';
-import content from "../data/path-selection.json";
+import content from "../data/en/path-selection.json";
 import CardInfo from "./CardInfo";
 
 function PathSelection(props) {
@@ -16,16 +16,16 @@ function PathSelection(props) {
         <h2>{content.heading_two}</h2>
 
         <Grid row gap>
-            <CardInfo 
-                header={content.update_btn_header} 
-                paragraph={content.update_btn_paragraph} 
+            <CardInfo
+                header={content.update_btn_header}
+                paragraph={content.update_btn_paragraph}
                 button={content.update_btn_txt}
                 role={"button"}
                 onClick={() => {props.getRegPath("update"), props.handleNext()}}>
             </CardInfo>
-            <CardInfo 
-                header={content.new_btn_header} 
-                paragraph={content.new_btn_paragraph} 
+            <CardInfo
+                header={content.new_btn_header}
+                paragraph={content.new_btn_paragraph}
                 button={content.new_btn_txt}
                 role={"button"}
                 onClick={() => {props.getRegPath("new"),  props.handleNext()}}>
