+++ src/components/RegTypes/EligibilityCheckboxes.jsx
@@ -1,5 +1,5 @@
 import { Button, Link, Fieldset, Checkbox, Label } from '@trussworks/react-uswds';
-import data from "../../data/state-selection.json";
+import data from "../../data/en/state-selection.json";
 import reactStringReplace from 'react-string-replace';
 
 function EligibilityCheckboxes(props) {
@@ -13,7 +13,7 @@ function EligibilityCheckboxes(props) {
         {content.download_form_link.replace("%state_name%", props.stateName)}
     </Link>
     );
-      
+
     return (
         <>
         <form onSubmit={() => {props.handleNext()}}>
@@ -32,7 +32,7 @@ function EligibilityCheckboxes(props) {
                         required={true}
                         defaultChecked={props.checkboxes.citizen}
                         onChange={(e) => props.handleCheckbox(e.target.checked, 'citizen', 0)}
-                    />              
+                    />
                     <Checkbox
                         id="age"
                         name="eligibility-checkbox"
@@ -44,11 +44,11 @@ function EligibilityCheckboxes(props) {
                         onChange={(e) => props.handleCheckbox(e.target.checked, 'age', 1)}
                     />
                 </div>
-            {props.checkboxes.checkboxesValid && 
+            {props.checkboxes.checkboxesValid &&
                 <span id="eligibility-error" rol="alert" className='error-text'>
                     Both boxes must be checked to continue.
                 </span>
-            }  
+            }
             </div>
         </Fieldset>
